requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.6.1
torchmetrics>=1.5.2
numpy>=1.24.1
pandas>=2.0.3
scikit-learn>=1.3.2
scipy>=1.10.1

# Chemistry libraries
rdkit-pypi>=2022.9.5

# Graph libraries
dgl-cu111>=0.6.1

# Utilities
tqdm>=4.67.1
pyyaml>=6.0.2
colorlog>=6.9.0

# Optional: for hyperparameter optimization
optuna>=4.3.0

# Optional: for web interface
streamlit>=1.40.1

